% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CD64}
\alias{CD64}
\title{Example data 2: CD64 data}
\format{
A data frame with 27 rows (studies) and 11 variables:
\describe{
  \item{study}{study ID}
  \item{Author}{authors of studies}
  \item{TP}{True positive}
  \item{FN}{False negative}
  \item{FP}{False positive}
  \item{TN}{True negative}
  \item{cutoff}{Cutoff values}
  \item{unit}{Unit of cutoff}
  \item{unit.grp}{Unit group}
  \item{cutoff.grp}{Cutoff group}
  \item{subgroup}{Subgroup}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/22940278/}
}
\usage{
CD64
}
\description{
{``To comprehensively and quantitatively summarize the accuracy of neutrophil CD64 in the early diagnosis of bacterial infection.''}
}
\keyword{datasets}
