% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CD64_2}
\alias{CD64_2}
\title{CD64 data after logit transformation}
\format{
A data frame with 27 rows and 6 variables:
\describe{
  \item{sens}{Sensitivity}
  \item{spec}{Sepcificity}
  \item{y1}{logit transformed sensitivity}
  \item{y2}{logit transformed specificity}
  \item{v1}{variance of logit transformed sensitivity}
  \item{v2}{variance of logit transformed specificity}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/22940278/}
}
\usage{
CD64_2
}
\description{
Logit-transformed CD64 data (after continuity correction)
}
\seealso{
\code{\link{CD64}}
}
\keyword{datasets}
