% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.pre.R
\name{correction}
\alias{correction}
\title{Continuity correction}
\usage{
correction(data, value = 0.5, type = c("single", "all"))
}
\arguments{
\item{data}{data}

\item{value}{value}

\item{type}{type}
}
\value{
data
}
\description{
Continuity correction
}
\examples{
correction(IVD)

}
