% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dta}
\alias{dta}
\title{Manual DTA data}
\format{
A data frame with 20 rows and 7 variables:
\describe{
  \item{se}{}
  \item{sp}{}
  \item{y1}{}
  \item{y2}{}
  \item{v1}{}
  \item{v2}{}
  \item{ldor.t}{}
}
}
\usage{
dta
}
\description{
Simulated DTA data
}
\keyword{datasets}
