% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtametasa.fc.R
\name{dtametasa.fc}
\alias{dtametasa.fc}
\title{Sensitivity analysis of diagnostic meta-analysis with prespecified contrast vector}
\usage{
dtametasa.fc(
  data,
  p,
  c1.square = 0.5,
  beta0 = 1,
  reitsma.par0 = NULL,
  beta.interval = c(0, 2),
  alpha.interval = c(-3, 3),
  ci.level = 0.95,
  correct.value = 0.5,
  correct.type = c("single", "all")[1],
  show.warn.message = FALSE,
  alpha.root.extendInt = "downX",
  eps = sqrt(.Machine$double.eps),
  sauc.type = c("sroc", "hsroc")[1],
  ...
)
}
\arguments{
\item{data}{Data with variable names either
\{TP, FN, TN, FP\} or \{y1, y2, v1, v2\}.
If not, please change the variable names.
Either data with the number of subjects or logit transformed data works.}

\item{p}{Specified probability of selection (or publication); Pr(select) = p}

\item{c1.square}{Pre-specified \eqn{c_1^2}{c1-square}.
\eqn{c_2^2 = 1 - c_1^2}{c2-square = 1-c1-square}}

\item{beta0}{An initial value of \eqn{\beta}{b}.
Avoid to start from 0.
Bad initial value will cause non-convergence results.}

\item{reitsma.par0}{Initial values used for estimating the parameters in the bivariate random effects model (Reitsma's model).
It should be changed by a vector of \code{c(mu1, mu2, tau1, tau2, rho)}.
Bad initial values will cause non-convergence results.
Default is \code{NULL}, which uses estimated results from Reitsma's model without taking into account the publication bias.}

\item{beta.interval}{The constraint interval for \eqn{\beta}{beta}.
The estimation of \eqn{\beta}{b} will be searched within the interval.
Take either the positive interval (>0) or the negative interval (<0).}

\item{alpha.interval}{The constraint interval for \eqn{\alpha}{alpha}.
then, the root of \eqn{\alpha}{a} will be searched within the interval.}

\item{ci.level}{The significant value for confidence interval of SAUC.s}

\item{correct.value}{Imputation value for ``continuity correction''.}

\item{correct.type}{Two types of ``continuity correction''.
\code{"single"}: input value for single study.
\code{"all"}: input value for all the cells.}

\item{show.warn.message}{Whether to show the warning messages.}

\item{alpha.root.extendInt}{See \code{extendInt} augment in function \code{\link[stats]{uniroot}}.}

\item{eps}{A minimum positive value.}

\item{sauc.type}{Two types of SAUC values.
Area under the Reitsma's SROC curve(\code{"sroc"}) or under the Rutter's HSROC curve (\code{"hsroc"}).}

\item{...}{See other augments in function \code{\link[stats]{uniroot}}.}
}
\value{
confidence interval,
convergence list,
logit transformed data
}
\description{
Sensitivity analysis with pre-specified (fixed) contrast vector \eqn{\boldsymbol{c} = (c_1, c_2)}
}
\details{
\describe{
\item{Continuity correction}{\url{https://en.wikipedia.org/wiki/Continuity_correction}}
}
}
\examples{

sa.fit1 <- dtametasa.fc(IVD, p = 0.7)
sa.fit1

sa.fit2 <- dtametasa.fc(IVD, p = 0.7, sauc.type = "sroc", correct.type = "all")
sa.fit2

sa.fit3 <- dtametasa.fc(IVD, p = 0.7, sauc.type = "hsroc")
sa.fit3

}
\seealso{
\code{\link[stats]{uniroot}},
\code{\link{dtametasa.rc}}.
}
