% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtasens1.R
\name{dtasens1}
\alias{dtasens1}
\title{dtasens1 function}
\usage{
dtasens1(
  data,
  p,
  c1 = sqrt(1),
  start6 = NULL,
  b.wid = c(0, 1),
  c.value = 0.5,
  c.type = "single",
  opt.type = c("optim", "nlminb"),
  warn = FALSE,
  show.data = FALSE,
  p.hat = FALSE,
  auc.all = FALSE,
  interval = c(-100, 100),
  ...
)
}
\arguments{
\item{data}{data}

\item{p}{p}

\item{c1}{c1}

\item{start6}{par}

\item{b.wid}{par}

\item{c.value}{par}

\item{c.type}{par}

\item{opt.type}{par}

\item{warn}{par}

\item{show.data}{par}

\item{p.hat}{par}

\item{auc.all}{par}

\item{interval}{par}

\item{...}{par}
}
\value{
convergence list
}
\description{
dtasens1 function
}
\examples{
dtasens1(dta, p = 0.7)

}
