% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtasens2.R
\name{dtasens2}
\alias{dtasens2}
\title{dtasens2 function}
\usage{
dtasens2(
  data,
  p,
  start7 = NULL,
  b.wid = c(0, 1),
  c.value = 0.5,
  c.type = "single",
  opt.type = c("optim", "nlminb"),
  warn = FALSE,
  show.data = FALSE,
  p.hat = FALSE,
  auc.all = FALSE,
  interval = c(-100, 100),
  ...
)
}
\arguments{
\item{data}{data}

\item{p}{p}

\item{start7}{par}

\item{b.wid}{par}

\item{c.value}{par}

\item{c.type}{par}

\item{opt.type}{par}

\item{warn}{par}

\item{show.data}{par}

\item{p.hat}{par}

\item{auc.all}{par}

\item{interval}{par}

\item{...}{par}
}
\value{
convergence list
}
\description{
dtasens2 function
}
\examples{
dtasens2(dta, p = 0.7)

}
