% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IVD}
\alias{IVD}
\title{IVD data}
\format{
A data frame with 33 rows and 5 variables:
\describe{
  \item{study}{}
  \item{TP}{}
  \item{FN}{}
  \item{FP}{}
  \item{TN}{}
}
}
\usage{
IVD
}
\description{
IVD data
}
\keyword{datasets}
