% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IVD_2}
\alias{IVD_2}
\title{IVD data after logit transformation}
\format{
A data frame with 33 rows and 6 variables:
\describe{
  \item{sens}{Sensitivity}
  \item{spec}{Sepcificity}
  \item{y1}{logit transformed sensitivity}
  \item{y2}{logit transformed specificity}
  \item{v1}{variance of logit transformed sensitivity}
  \item{v2}{variance of logit transformed specificity}
}
}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/9315770}
}
\usage{
IVD_2
}
\description{
{Logit-transformed IVD data (after continuity correction)}
}
\seealso{
\code{\link{IVD}}
}
\keyword{datasets}
