% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAUC.R
\name{mAUC.ci}
\alias{mAUC.ci}
\title{Calculate multiple AUC with CI}
\usage{
mAUC.ci(par.matrix, ci.level = 0.95)
}
\arguments{
\item{par.matrix}{cbind(u1, u2, t1, t2, r)}

\item{ci.level}{ci.level}
}
\value{
auc
}
\description{
Calculate multiple AUC.ci
}
\examples{
par.matrix <-matrix(c(1,1,0.5, 0.5, -0.6, 1,1,1, 2, -0.6), 5,2)
mAUC.ci(par.matrix, ci.level = 0.9)

}
