% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAUC.R
\name{mAUC}
\alias{mAUC}
\title{Calculate multiple AUC}
\usage{
mAUC(par.matrix)
}
\arguments{
\item{par.matrix}{cbind(u1, u2, t1, t2, r)}
}
\value{
auc
}
\description{
Calculate multiple AUC
}
\examples{

par.matrix <-matrix(c(1,1,0.5, 0.5, -0.6, 1,1,1, 2, -0.6), 5,2)
mAUC(par.matrix)

}
