% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sROC.R
\name{mROC}
\alias{mROC}
\title{Multiple ROC curves}
\usage{
mROC(
  par.matrix,
  s.point = TRUE,
  s.line = FALSE,
  new.plot = TRUE,
  legend = FALSE,
  p.vec,
  cols = NULL,
  ...
)
}
\arguments{
\item{par.matrix}{cbind(u1, u2, t1, t2, r)}

\item{s.point}{s.point}

\item{s.line}{s.line}

\item{new.plot}{new.plot}

\item{legend}{legend}

\item{p.vec}{p.vec}

\item{cols}{set cols}

\item{...}{par}
}
\value{
plot
}
\description{
Multiple ROC curves
}
\examples{
par.matrix <-matrix(c(1,1,0.5, 0.5, -0.6, 1,1,1, 2, -0.6), 5,2)
mROC(par.matrix, legend = TRUE, p.vec = c(0.9, 0.5), cols = 1:2)

}
