% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim2.R
\name{optim2}
\alias{optim2}
\title{optim2 function}
\usage{
optim2(
  data,
  p,
  start7 = c(0, 0, 1, 1, -0.5, 0, 0.7),
  b.wid = c(0, 1),
  c.value = 0.5,
  c.type = "single",
  opt.type = c("optim", "nlminb"),
  warn = FALSE,
  show.data = FALSE,
  p.hat = FALSE,
  auc.all = FALSE,
  interval = c(-100, 100),
  ...
)
}
\arguments{
\item{data}{data}

\item{p}{p}

\item{start7}{par}

\item{b.wid}{par}

\item{c.value}{par}

\item{c.type}{par}

\item{opt.type}{par}

\item{warn}{par}

\item{show.data}{par}

\item{p.hat}{par}

\item{auc.all}{par}

\item{interval}{par}

\item{...}{par}
}
\value{
convergence list
}
\description{
optim2 function
}
\examples{
optim2(dta, p = 0.7)

}
