% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAUC.ci.R
\name{plot.sAUC.ci}
\alias{plot.sAUC.ci}
\title{Plot sROC.ci results}
\usage{
\method{plot}{sAUC.ci}(
  x,
  roc.ci.col = "grey",
  roc.ci.lty = 1,
  roc.ci.lwd = 1,
  add.sROC.ci = TRUE,
  ci.level = 0.95,
  ...
)
}
\arguments{
\item{x}{object from function \code{\link{sAUC.ci}}.}

\item{roc.ci.col}{The color of the CIs of sROC.
Default is grey.}

\item{roc.ci.lty}{The line type of CIs.
Default is solid line.}

\item{roc.ci.lwd}{The line width of CIs.
Defalt is 1.}

\item{add.sROC.ci}{Whether to add the plot of sROC with the CIs lines
onto a new plot.
Default is not to add.}

\item{ci.level}{The significant value for confidence interval.
Default is 0.95, hence, a 2-tailed 95% CIs will be calculated by
profile likelihood.}

\item{...}{Other augments in function \code{\link{sROC}}}
}
\description{
Plot the sROC from function \code{\link{sAUC.ci}}
}
\examples{

sa1 <- dtametasa.fc(IVD, p = 0.7)

sROC(sa1)
ci <- sAUC.ci(sa1, B = 5, set.seed = 1, plot.ROC.ci = FALSE)

plot(ci)
p <- plot(ci)

}
\seealso{
\code{\link{sAUC.ci}};
\code{\link[graphics]{matplot}},
}
