% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dtametasa.R
\name{print.dtametasa}
\alias{print.dtametasa}
\title{Print dtametasa.fc and dtametasa.rc functions results}
\usage{
\method{print}{dtametasa}(x, digits = 3, ...)
}
\arguments{
\item{x}{object from function \code{dtametasa.fc} or \code{dtametasa.rc}}

\item{digits}{digits of the results}

\item{...}{other parameters in function \code{\link{print}}}
}
\description{
Print results from function \code{\link{dtametasa.fc}} or \code{\link{dtametasa.rc}}
}
\seealso{
\code{\link{dtametasa.fc}},
\code{\link{dtametasa.rc}},
\code{\link[base]{print}}.
}
