% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dtasens.R
\name{print.DTAsens}
\alias{print.DTAsens}
\title{Print DTAsens results}
\usage{
\method{print}{DTAsens}(object, digits = 3, ...)
}
\arguments{
\item{object}{object}

\item{digits}{object}

\item{...}{par}
}
\value{
par
}
\description{
Print DTAsens results
}
\examples{
(opt1 <- dtasens1(IVD, p = 0.7))
(opt2 <- dtasens2(IVD, p = 0.7))
}
