% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAUC.ci.R
\name{print.sAUC.ci}
\alias{print.sAUC.ci}
\title{Print sAUC CI}
\usage{
\method{print}{sAUC.ci}(x, digits = 3, ...)
}
\arguments{
\item{x}{object}

\item{digits}{digits}

\item{...}{graphical parameters to plot}
}
\description{
Print sAUC CI
}
