% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sauc.ci.R
\name{sAUC.ci}
\alias{sAUC.ci}
\title{Parametric bootstrapped confidence interval of summary AUC}
\usage{
sAUC.ci(
  object,
  B = 1000,
  ncores = 0,
  type = "SOCK",
  ci.level = 0.95,
  set.seed = NULL,
  hide.progress = FALSE
)
}
\arguments{
\item{object}{object from function \code{dtametasa.fc} or \code{dtametasa.rc}.}

\item{B}{The times for parametric bootstrapping.
Default is 1000.
It may be time consuming.}

\item{ncores}{Set the an integer number of cores that will be used in parallel computing.
Default is 0, which detect the number of cores in the local device.}

\item{type}{Computing types in the parallel computing.
See \code{type} augment in function \code{\link[parallel]{makeCluster}}.}

\item{ci.level}{The significant value for confidence interval.
Default is 0.95, hence, a 2-tailed 95% CIs will be calculated by
profile likelihood.}

\item{set.seed}{Give a integer to set the seed.
Default is NULL, which does not set any seed to fix the bootstrap results.}

\item{hide.progress}{Whether to hide the progress bar in the calculation.
Default is not to hide.}
}
\value{
CI of sAUC
}
\description{
Calculate the parametric bootstrap confidence interval (CI) for sAUC
}
\examples{

## Here, we set B = 5 to save time in the calculation.
## But, the results are not reliable.

sa1 <- dtametasa.fc(IVD, p = 0.7)
(ci <- sAUC.ci(sa1, B = 5, set.seed = 1))

}
\seealso{
\code{\link[parallel]{makeCluster}},
}
