% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAUC.R
\name{sAUC}
\alias{sAUC}
\title{Calculate Single AUC}
\usage{
sAUC(object)
}
\arguments{
\item{object}{c(u1, u2, t1, t2, r)}
}
\value{
integrate
}
\description{
Calculate single AUC
}
\examples{

par.vec <- c(1,1, 0.5, 0.5, -0.7)
sAUC(par.vec)

opt1 <- dtasens1(IVD, p = 0.7)
sAUC(opt1)
}
