% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.data.R
\docType{data}
\name{Scheidler}
\alias{Scheidler}
\title{Example data 2: Scheidler data}
\format{
A data frame with 44 rows and 5 variables:
\describe{
  \item{ID}{study ID}
  \item{author}{authors of studies}
  \item{year}{publication years}
  \item{method}{1 is CT, 2 is LAG, 3 is MR}
  \item{TP}{True positive}
  \item{FN}{False negative}
  \item{FP}{False positive}
  \item{TN}{True negative}
}
}
\source{
\url{http://annals.org/article.aspx?doi=10.7326/0003-4819-142-6-200503150-00011}
}
\usage{
Scheidler
}
\description{
{``To apply meta-analysis to compare the utility of
lymphangiography (LAG), computed tomography (CT), and magnetic resonance (MR) imaging
for the diagnosis of lymph node metastasis in patients with cervical cancer.''}
}
\keyword{datasets}
