% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.data.R
\docType{data}
\name{Scheidler2}
\alias{Scheidler2}
\title{Scheidler data after logit transformation}
\format{
A data frame with 44 rows and 7 variables:
\describe{
  \item{sens}{Sensitivity}
  \item{spec}{Sepcificity}
  \item{y1}{logit transformed sensitivity}
  \item{y2}{logit transformed specificity}
  \item{v1}{variance of logit transformed sensitivity}
  \item{v2}{variance of logit transformed specificity}
  \item{ldor.t}{t-statistic of log diagnostic odds ratio}
}
}
\source{
\url{http://annals.org/article.aspx?doi=10.7326/0003-4819-142-6-200503150-00011}
}
\usage{
Scheidler2
}
\description{
Logit-transformed Scheidler data (after continuity correction)
}
\seealso{
\code{\link{Scheidler}}
}
\keyword{datasets}
