% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.pdata.R
\name{sim.pdata}
\alias{sim.pdata}
\title{Simulate the population diagnostic data}
\usage{
sim.pdata(par)
}
\arguments{
\item{par}{A vector of c(S, u1, u2, t11, t22, t12, c11, c22)}
}
\value{
a simulation of logit transformed diagnostic study data
}
\description{
Simulate the population diagnostic data,
based on bivariate random effects model.
Within-study variances are distribution as normal distribution
with mean 0.5 and variance 0.25
}
\examples{

par <- c(20, 0, 0, 1, 1, -0.3, 0.5, 0.5)
pdata <- sim.pdata(par)


}
\seealso{
\code{\link{dtametasa.fc}},
\code{\link{dtametasa.rc}}.
}
