% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sroc.R
\name{sROC.matrix}
\alias{sROC.matrix}
\title{Plot summary ROC curves}
\usage{
sROC.matrix(
  par,
  add = FALSE,
  ncols = NULL,
  roc.lty = 1,
  roc.lwd = 1,
  add.sum.point = TRUE,
  legend = FALSE,
  p.vec,
  legend.text = paste0("p = ", p.vec),
  legend.cex = 1,
  spoint.pch = 19,
  spoint.cex = 1,
  xlab = "1 - specificity",
  ylab = "Sensitivity",
  ...
)
}
\arguments{
\item{par}{It can be a vector of (u1 u2 t22 t12),
or a matrix with 4 rows.
Each column is the vector \code{c(u1, u2, t22, t12)}.}

\item{add}{Whether to add the plot into an existed plot.
Default is \code{FALSE}, to create a new plot.}

\item{ncols}{Set different colors for multiple sROC.
Defult is \code{NULL}, that uses different grey's colors.}

\item{roc.lty}{The line tyoe of sROC.
Default is solid lines.}

\item{roc.lwd}{The line width of sROC.
Default is 1.}

\item{add.sum.point}{Whether to add the summary point in the sROC plot.
Default it not the add.}

\item{legend}{Whether to add legend into the plot.
Default is not to add.}

\item{p.vec}{If add the legend (\code{legend = TRUE}),
define the probability sequence.}

\item{legend.text}{If add the legend (\code{legend = TRUE}),
define the legend context.}

\item{legend.cex}{The font size of legend.}

\item{spoint.pch}{The point type of the summary point in sROC.}

\item{spoint.cex}{The point size of the summary point in sROC.}

\item{xlab}{The label of x-axis.
Default is: 1-specificity.}

\item{ylab}{The label of y-axis.
Default is Sensitivity.}

\item{...}{Other augments in function \code{\link{points}} or function \code{\link{curve}}}
}
\value{
sROC plot
}
\description{
Plot multiple ROC curves
}
\examples{

# Use matrix

par.matrix <-cbind(c(1, 1.5, 0.5, -0.2),
                   c(1, 1, 1, -0.6),
                   c(1.8, 2, 1,-0.6))

sROC.matrix(par.matrix, legend = TRUE, p.vec = seq(0.2,0.6,0.2), legend.cex = 0.9)

sROC.matrix(par.matrix, legend = TRUE,
      legend.text = c("l1", "l2", "l3"),
      legend.cex = 0.9, ncols = 1:3)

# Use vector

par.vec <- c(1,1.5,0.5,-0.2)

sROC.matrix(par.vec)

}
\seealso{
\code{\link[graphics]{points}},
\code{\link[graphics]{curve}},
\code{\link{dtametasa.fc}},
\code{\link{dtametasa.rc}}.
}
