% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sROC_AUC.R
\name{sROC}
\alias{sROC}
\title{ROC plot}
\usage{
sROC(u1, u2, t22, t12, par = NULL, add = FALSE, s.point = TRUE, ...)
}
\arguments{
\item{u1}{u1}

\item{u2}{u2}

\item{t22}{t22}

\item{t12}{t12}

\item{par}{c(u1, u2, t22, t12)}

\item{add}{par}

\item{s.point}{par}

\item{...}{par}
}
\value{
plot
}
\description{
ROC plot
}
\examples{
sROC(1,1,0.5, -0.5)

}
