% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtametasa.fc.R
\name{dtametasa.fc}
\alias{dtametasa.fc}
\title{Sensitivity analysis of diagnostic meta-analysis with fixed c}
\usage{
dtametasa.fc(
  data,
  p,
  c1.sq = 0.5,
  correct.value = 0.5,
  correct.type = "all",
  brem.init = NULL,
  b.init = 1,
  b.interval = c(0, 2),
  a.interval = c(-3, 3),
  positive.r = TRUE,
  ci.level = 0.95,
  show.warn.message = FALSE,
  a.root.extendInt = "downX",
  ...
)
}
\arguments{
\item{data}{Data with variable names either
\{TP, FN, TN, FP\} or \{y1, y2, v1, v2\}.
If not, please change the variable names.
Either data with the number of subjects or logit transformed data works.}

\item{p}{Specified probability of selection (or publication); Pr(select) = p}

\item{c1.sq}{Pre-specified \eqn{c_1^2}{c1-square}.
Default value is 0.5;
hence, \eqn{c_2^2 = 1 - c_1^2}{c2-square = 1-c1-square}}

\item{correct.value}{Imputation value for ``continuity correction''.
Default is 0.5;
When zero cell exists, add some value.}

\item{correct.type}{2 ways for ``continuity correction''.
Default is \code{all}: add \code{correct.value} into all the cells of the row with zero value;
if \code{single}: add \code{correct.value} into only the cells with zero value.}

\item{brem.init}{Initial values used for estimating the parameters in the bivariate random effects model.
Default is \code{NA}, which set the estimation without considering publication bias as initial values;
It should be changed by a vector of u1 u2 t1 t2 r, \code{c(u1, u2, t1, t2, r)}.
Bad initial values will cause non-convergence results.}

\item{b.init}{An initial value of \eqn{\beta}{b}.
Default is 0.1.
Avoid to start from 0.
Bad initial value will cause non-convergence results.}

\item{b.interval}{The constraint interval for \eqn{\beta}{b}.
Default is in the interval \code{c(0,2)}.
Positive side (>0) is recommended to be adopted.
The estimation of \eqn{\beta}{b} will be searched within the interval.}

\item{a.interval}{The constraint interval for \eqn{\alpha}{a}.
Default is in the interval \code{c(-3,3)}.
then, the root of \eqn{\alpha}{a} will be searched within the interval.}

\item{positive.r}{Whether to permit correlation parameter \eqn{\rho}{r} to be a positive value,
though it is often a negative value
Default \code{TRUE}, to permit positive, in which \eqn{\rho}{r} is in [-1, 1].
If \code{FALSE}, \eqn{\rho}{r} is in [-1, 0)}

\item{ci.level}{The significant value for confidence interval.
Default is 0.95, hence, a 2-tailed 95% confidence interval will be calculated by
profile likelihood.}

\item{show.warn.message}{Whether to show the warning messages.
Default is to hide warning messages.}

\item{a.root.extendInt}{See \code{extendInt} augment in function \code{\link[stats]{uniroot}}.}

\item{...}{See other augments in function \code{\link[stats]{uniroot}}.}
}
\value{
confidence interval,
convergence list,
logit transformed data
}
\description{
DTA sensitivity analysis with pre-specified (fixed) c1 and c2
}
\details{
\describe{
\item{Continuity correction}{\url{https://en.wikipedia.org/wiki/Continuity_correction}}
}
}
\examples{

sa1 <- dtametasa.fc(IVD, p = 0.7)
sa1

}
\seealso{
\code{\link[stats]{uniroot}},
\code{\link{dtametasa.rc}}.
}
