% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtasens1.R
\name{dtasens1}
\alias{dtasens1}
\title{dtasens1 function}
\usage{
dtasens1(
  data,
  p,
  start6 = NULL,
  c1 = sqrt(0.5),
  correct.value = 0.5,
  correct.type = "single",
  optimize.type = c("optim", "nlminb"),
  show.warn.message = FALSE,
  show.data = FALSE,
  show.p.hat = FALSE,
  show.auc.all = FALSE,
  b.interval = c(0, 2),
  a.interval = c(-10, 0),
  a.root.extendInt = "downX",
  ...
)
}
\arguments{
\item{data}{data}

\item{p}{p}

\item{start6}{par}

\item{c1}{c1}

\item{correct.value}{par}

\item{correct.type}{par}

\item{optimize.type}{par}

\item{show.warn.message}{par}

\item{show.data}{par}

\item{show.p.hat}{par}

\item{show.auc.all}{par}

\item{b.interval}{par}

\item{a.interval}{par}

\item{a.root.extendInt}{par}

\item{...}{par}
}
\value{
convergence list
}
\description{
dtasens1 function
}
\examples{
dtasens1(dta, p = 0.7)

}
