% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtasens1.R
\name{dtasens1}
\alias{dtasens1}
\title{dtasens1 function}
\usage{
dtasens1(
  data,
  p,
  c1 = sqrt(0.5),
  correct.value = 0.5,
  correct.type = "single",
  start5 = NULL,
  b0 = 0.1,
  b.interval = c(0, 2),
  a.interval = c(-5, 3),
  pos.r = TRUE,
  ci.level = 0.95,
  show.warn.message = FALSE,
  a.root.extendInt = "downX",
  ...
)
}
\arguments{
\item{data}{data}

\item{p}{p}

\item{c1}{c1}

\item{correct.value}{par}

\item{correct.type}{"single", "all"}

\item{start5}{A vector of u1 u2 t1 t2 r}

\item{b0}{A start value of b}

\item{b.interval}{par}

\item{a.interval}{par}

\item{pos.r}{permit positive r}

\item{ci.level}{0.95}

\item{show.warn.message}{par}

\item{a.root.extendInt}{par}

\item{...}{par}
}
\value{
convergence list
}
\description{
dtasens1 function
}
\examples{

opt1 <- dtasens1(IVD, p = 0.7)

}
