% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtasens2.R
\name{dtasens2}
\alias{dtasens2}
\title{DTAsens function 2}
\usage{
dtasens2(
  data,
  p,
  correct.value = 0.5,
  correct.type = "all",
  start5 = NULL,
  b0 = 0.1,
  c10 = sqrt(0.5),
  b.interval = c(0, 2),
  a.interval = c(-5, 3),
  pos.r = TRUE,
  ci.level = 0.95,
  show.warn.message = FALSE,
  a.root.extendInt = "downX",
  ...
)
}
\arguments{
\item{data}{Data with variable names either
\{TP, FN, TN, FP\} or \{y1, y2, v1, v2\}.
If not, please change the variable names.
Either data with the number of subjects or logit transformed data works.}

\item{p}{Specified probability of selection (or publication); Pr(select) = p}

\item{correct.value}{Imputation value for ``continuity correction''.
Default is 0.5;
When zero cell exists, add some value.}

\item{correct.type}{2 ways for ``continuity correction''.
Default is \code{all}: add \code{correct.value} into all the cells of the row with zero value;
if \code{single}: add \code{correct.value} into only the cells with zero value.}

\item{start5}{Starting values used for estimating the parameters.
Default is \code{NULL}, which set the estimation from standard model as starting values;
It should be changed by a vector of u1 u2 t1 t2 r, \code{c(u1, u2, t1, t2, r)}.
Bad starting values will cause non-convergence results.}

\item{b0}{A starting value of \eqn{\beta}{b}. 
Default is 0.1.
Avoid to start from 0.
Bad starting value will cause non-convergence results.}

\item{c10}{A starting value of \eqn{c_1}{c1}. 
Default is \eqn{\sqrt{0.5}}{0.701}.
Avoid to start from 0.
Bad starting value will cause non-convergence results.}

\item{b.interval}{The constraint interval for \eqn{\beta}{b}.
Default is in the interval \code{c(0,2)}.
Positive side (>0) is recommended to be adopted.
The estimation of \eqn{\beta}{b} will be searched within the interval.}

\item{a.interval}{The constraint interval for \eqn{\alpha}{a}.
Default is in the interval \code{c(-5,3)}.
then, the root of \eqn{\alpha}{a} will be searched within the interval.}

\item{pos.r}{Whether to permit correlation parameter \eqn{\rho}{r} to be a positive value,
though it is often a negative value
Default is to permit positive.}

\item{ci.level}{The significant value for confidence interval.
Default is 0.95, hence, a 2-tailed 95% confidence interval will be calculated by
profile likelihood.}

\item{show.warn.message}{Whether to show the warning messages.
Default is to hide warning messages.}

\item{a.root.extendInt}{See \code{extendInt} augments in function \code{\link[stats]{uniroot}}.}

\item{...}{See other augments in function \code{\link[stats]{uniroot}}.}
}
\value{
confidence interval,
convergence list,
logit transformed data
}
\description{
DTA sensitivity analysis without pre-specified c1 and c2
}
\details{
\describe{
\item{Continuity correction}{\url{https://en.wikipedia.org/wiki/Continuity_correction}}
}
}
\examples{

opt2 <- dtasens2(IVD, p = 0.7)
opt2

}
\seealso{
\code{\link[stats]{uniroot}},
\code{\link{dtasens1}}.
}
