% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtasens2.R
\name{dtasens2}
\alias{dtasens2}
\title{dtasens2 function}
\usage{
dtasens2(
  data,
  p,
  start7 = NULL,
  correct.value = 0.5,
  correct.type = "single",
  optimize.type = c("optim", "nlminb"),
  show.warn.message = FALSE,
  show.data = FALSE,
  show.p.hat = FALSE,
  show.auc.all = FALSE,
  b.interval = c(0, 2),
  a.interval = c(-10, 0),
  a.root.extendInt = "downX",
  ...
)
}
\arguments{
\item{data}{data}

\item{p}{p}

\item{start7}{par}

\item{correct.value}{par}

\item{correct.type}{par}

\item{optimize.type}{par}

\item{show.warn.message}{par}

\item{show.data}{par}

\item{show.p.hat}{par}

\item{show.auc.all}{par}

\item{b.interval}{par}

\item{a.interval}{par}

\item{a.root.extendInt}{par}

\item{...}{par}
}
\value{
convergence list
}
\description{
dtasens2 function
}
\examples{
dtasens2(dta, p = 0.7)

}
