% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtasens2.R
\name{dtasens2}
\alias{dtasens2}
\title{dtasens2 function}
\usage{
dtasens2(
  data,
  p,
  correct.value = 0.5,
  correct.type = "single",
  start5 = NULL,
  b0 = 0,
  c0 = sqrt(0.5),
  b.interval = c(0, 3),
  a.interval = c(-5, 5),
  optimize.type = c("optim", "nlminb"),
  pos.r = FALSE,
  show.warn.message = FALSE,
  show.data = FALSE,
  show.p.hat = FALSE,
  show.auc.all = FALSE,
  a.root.extendInt = "yes",
  ...
)
}
\arguments{
\item{data}{data}

\item{p}{p}

\item{correct.value}{par}

\item{correct.type}{par}

\item{start5}{a vector of u1 u2 t1 t2 r}

\item{b0}{b}

\item{c0}{c1}

\item{b.interval}{par one-side}

\item{a.interval}{par}

\item{optimize.type}{par}

\item{pos.r}{permit positive r}

\item{show.warn.message}{par}

\item{show.data}{par}

\item{show.p.hat}{par}

\item{show.auc.all}{par}

\item{a.root.extendInt}{par}

\item{...}{par}
}
\value{
convergence list
}
\description{
dtasens2 function
}
\examples{
dtasens2(dta, p = 0.7)

}
