% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtasens2.R
\name{dtasens2}
\alias{dtasens2}
\title{dtasens2 function}
\usage{
dtasens2(
  data,
  p,
  correct.value = 0.5,
  correct.type = "single",
  start5 = NULL,
  b0 = 0.1,
  c10 = sqrt(0.5),
  b.interval = c(0, 2),
  a.interval = c(-5, 3),
  pos.r = TRUE,
  ci.level = 0.95,
  show.warn.message = FALSE,
  a.root.extendInt = "downX",
  ...
)
}
\arguments{
\item{data}{data}

\item{p}{p}

\item{correct.value}{par}

\item{correct.type}{par}

\item{start5}{a vector of u1 u2 t1 t2 r}

\item{b0}{b}

\item{c10}{c1}

\item{b.interval}{par one-side}

\item{a.interval}{par}

\item{pos.r}{permit positive r}

\item{ci.level}{ci.level}

\item{show.warn.message}{par}

\item{a.root.extendInt}{par}

\item{...}{...}
}
\value{
convergence list
}
\description{
dtasens2 function
}
\examples{

opt1 <- dtasens1(IVD, p = 0.7)

}
