% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IVD}
\alias{IVD}
\title{Example data 1: IVD data}
\format{
A data frame with 33 rows (studies) and 8 variables:
\describe{
  \item{study}{study ID}
  \item{TP}{True positive}
  \item{FN}{False negative}
  \item{FP}{False positive}
  \item{TN}{True negative}
  \item{cutoff.grp}{Cutoff value by group}
  \item{type}{Diagnostic type}
  \item{cutoff}{Cutoff information}
  #' }
}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/9315770}
}
\usage{
IVD
}
\description{
{``To identify the most accurate methods for diagnosis of IVD-related bloodstream infection.''}
}
\keyword{datasets}
