% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sROC.R
\name{mROC}
\alias{mROC}
\title{Multiple ROC curves}
\usage{
mROC(
  par.matrix,
  add = FALSE,
  ncols = NULL,
  roc.lty = 1,
  roc.lwd = 1,
  add.sum.point = TRUE,
  legend = FALSE,
  p.vec,
  legend.text = paste0("p = ", p.vec),
  legend.cex = 1,
  spoint.pch = 19,
  spoint.cex = 1,
  ...
)
}
\arguments{
\item{par.matrix}{cbind(u1, u2, t1, t2, r)}

\item{add}{par}

\item{ncols}{ncols}

\item{roc.lty}{roc.lty}

\item{roc.lwd}{par}

\item{add.sum.point}{par}

\item{legend}{par}

\item{p.vec}{par}

\item{legend.text}{par}

\item{legend.cex}{par}

\item{spoint.pch}{par}

\item{spoint.cex}{par}

\item{...}{par}
}
\value{
plot
}
\description{
Multiple ROC curves
}
\examples{

par.matrix <-cbind(c(1,1.5,0.5, 0.5, -0.2),c(1, 1, 1, 2, -0.6), c(1.8, 2, 1, 2, -0.6))
p.vec <- seq(0.2,0.6,0.2)
mROC(par.matrix, legend = TRUE, p.vec = p.vec, legend.cex = 0.9)
mROC(par.matrix, legend = TRUE, legend.text = c("l1", "l2", "l3"), legend.cex = 0.9, ncols = 1:3)

}
