% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dtasens.R
\name{print.dtasens}
\alias{print.dtasens}
\title{Print dtasens1 and dtasens2}
\usage{
\method{print}{dtasens}(x, digits = 3, ...)
}
\arguments{
\item{x}{object from function \code{dtasens1} or \code{dtasens2}}

\item{digits}{digits of the results}

\item{...}{other parameters in function \code{\link{print}}}
}
\description{
Print results from function \code{\link{dtasens1}} or \code{\link{dtasens2}}
}
\examples{

(opt1 <- dtasens1(IVD, p = 0.7))
(opt2 <- dtasens2(IVD, p = 0.7))

}
\seealso{
\code{\link{dtasens1}},
\code{\link{dtasens2}},
\code{\link[base]{print}}.
}
