% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dtasens.R
\name{print.DTAsens}
\alias{print.DTAsens}
\title{Print DTAsens}
\usage{
\method{print}{DTAsens}(x, digits = 3, ...)
}
\arguments{
\item{x}{object}

\item{digits}{digits}

\item{...}{graphical parameters to plot}
}
\description{
Print DTAsens
}
\examples{

(opt1 <- dtasens1(IVD, p = 0.7))
(opt2 <- dtasens2(IVD, p = 0.7))

}
