% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAUC.ci.R
\name{print.sAUC.ci}
\alias{print.sAUC.ci}
\title{Print sAUC.ci results}
\usage{
\method{print}{sAUC.ci}(x, digits = 3, ...)
}
\arguments{
\item{x}{object from function \code{\link{sAUC.ci}}.}

\item{digits}{digits of the results.}

\item{...}{other parameters in function \code{\link{print}}.}
}
\description{
Print results from function \code{\link{sAUC.ci}}
}
\seealso{
\code{\link{sAUC.ci}};
\code{\link[base]{print}}
}
