% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAUC.R
\name{sAUC.ci}
\alias{sAUC.ci}
\title{Calculate Single AUC with CI}
\usage{
sAUC.ci(par.vec, ci.level = 0.95)
}
\arguments{
\item{par.vec}{c(u1, u2, t1, t2, r)}

\item{ci.level}{ci.level}
}
\value{
integrate
}
\description{
Calculate single AUC.CI
}
\examples{
par.vec <- c(1,1, 0.5, 0.5, -0.6)
sAUC.ci(par.vec)

}
