% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAUC.ci.R
\name{sAUC.ci}
\alias{sAUC.ci}
\title{sAUC.ci function}
\usage{
sAUC.ci(
  object,
  B = 10,
  ncores,
  type = "SOCK",
  ci.level = 0.95,
  hide.progress = FALSE,
  plot.ROC.ci = FALSE,
  add.sum.point = FALSE,
  roc.ci.col = "grey",
  roc.ci.lty = 1,
  roc.ci.lwd = 1,
  ...
)
}
\arguments{
\item{object}{data}

\item{B}{p}

\item{ncores}{ncores}

\item{type}{c1}

\item{ci.level}{par}

\item{hide.progress}{par}

\item{plot.ROC.ci}{par}

\item{add.sum.point}{add.sum.point}

\item{roc.ci.col}{add.sum.point}

\item{roc.ci.lty}{roc.col}

\item{roc.ci.lwd}{roc.ci.lwd}

\item{...}{...}
}
\value{
convergence list
}
\description{
Parametric Bootstrap CI of sAUC
}
\examples{

opt1 <- dtasens1(IVD, p = 0.7)
(ci <- sAUC.ci(opt1, B = 5, plot.ROC.ci = FALSE))

}
