% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAUC.ci.R
\name{sAUC.ci}
\alias{sAUC.ci}
\title{sAUC.ci function}
\usage{
sAUC.ci(object, B = 10, ncores, type = "SOCK", ci.level = 0.95)
}
\arguments{
\item{object}{data}

\item{B}{p}

\item{ncores}{ncores}

\item{type}{c1}

\item{ci.level}{par}
}
\value{
convergence list
}
\description{
Parametric Bootstrap CI of sAUC
}
\examples{

opt1 <- dtasens1(IVD, p = 0.7)
(ci <- sAUC.ci(opt1))

}
