% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sauc.R
\name{sAUC}
\alias{sAUC}
\title{Calculate summary AUC}
\usage{
sAUC(par)
}
\arguments{
\item{par}{A vector of (u1 u2 t22 t12),
or a matrix with 4 rows.
Each column is the vector of \code{c(u1, u2, t22, t12)}.}
}
\value{
sAUC values
}
\description{
Calculate sAUC values from parameter vector or parameter matrix
}
\examples{

#Use matrix

par.matrix <-cbind(c(1,1,0.5, -0.6),
                   c(1,1,2, -0.6))

sAUC(par.matrix)

# Use vector

par.vec <- c(1,1,0.5, -0.6)

sAUC(par.vec)

}
