% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAUC.R
\name{sAUC}
\alias{sAUC}
\title{Calculate a single summary AUC value}
\usage{
sAUC(object)
}
\arguments{
\item{object}{The object from function \code{dtasens1} or \code{dtasens2};
or a vector of \code{c(u1, u2, t1, t2, r)}}
}
\value{
sAUC value
}
\description{
Calculate a single sAUC value
}
\examples{

opt1 <- dtasens1(IVD, p = 0.7)
sAUC(opt1)

par.vec <- c(1,1, 0.5, 0.5, -0.7)
sAUC(par.vec)

}
\seealso{
\code{\link[stats]{integrate}}
}
