% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sROC_AUC.R
\name{sROC.bunch}
\alias{sROC.bunch}
\title{ROC bunch plot}
\usage{
sROC.bunch(
  par.matrix,
  s.point = TRUE,
  s.line = FALSE,
  new.plot = TRUE,
  legend = TRUE,
  p.vec,
  ...
)
}
\arguments{
\item{par.matrix}{rbind(u1, u2, t22, t12)}

\item{s.point}{s.point}

\item{s.line}{s.line}

\item{new.plot}{new.plot}

\item{legend}{legend}

\item{p.vec}{p.vec}

\item{...}{par}
}
\value{
plot
}
\description{
ROC plot in bunch
}
\examples{
par.matrix <-matrix(c(1,1,0.5, -0.5,
                   1,1,0.5, -0.3), 4,2)
sROC.bunch(par.matrix, p.vec = c(0.9, 0.5))

}
