% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sroc.R
\name{sROC.matrix}
\alias{sROC.matrix}
\title{Plot summary ROC curves}
\usage{
sROC.matrix(
  par,
  add = FALSE,
  ncols = gray.colors(ncol(par), gamma = 1, start = 0.8, end = 0),
  sroc.lty = 1,
  sroc.lwd = 1,
  add.spoint = TRUE,
  spoint.pch = 18,
  spoint.cex = 2,
  xlab = "1 - specificity",
  ylab = "Sensitivity",
  ...
)
}
\arguments{
\item{par}{The object from function \code{dtametasa.fc} or \code{dtametasa.rc};
or a matrix with rows of \code{c(u1, u2, t1, t2, r)}}

\item{add}{Whether to add the plot into an existed plot.
Default is \code{FALSE}, to create a new plot.}

\item{ncols}{Set a vector of different colors for multiple sROC.
Defult uses different grey's colors.}

\item{sroc.lty}{The line tyoe of sROC.
Default is solid lines.}

\item{sroc.lwd}{The line width of sROC.
Default is 1.}

\item{add.spoint}{Whether to add the summary point in the sROC plot.
Default it not the add.}

\item{spoint.pch}{The point type of the summary point in sROC.}

\item{spoint.cex}{The point size of the summary point in sROC.}

\item{xlab}{The label of x-axis.
Default is: 1-specificity.}

\item{ylab}{The label of y-axis.
Default is Sensitivity.}

\item{...}{Other augments in function \code{\link{points}} or function \code{\link{curve}}}
}
\value{
sROC plot
}
\description{
Plot multiple ROC curves
}
\examples{
p.seq <- seq(0.5, 0.9, 0.1)
sa1.seq <- sapply(p.seq, function (p) dtametasa.fc(IVD, p)$par)
sROC.matrix(sa1.seq)




}
\seealso{
\code{\link[graphics]{points}},
\code{\link[graphics]{curve}},
\code{\link{dtametasa.fc}},
\code{\link{dtametasa.rc}}.
}
