% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sROC_AUC.R
\name{sROC}
\alias{sROC}
\title{ROC plot}
\usage{
sROC(par.vec = NULL, add = FALSE, s.point = TRUE, ...)
}
\arguments{
\item{par.vec}{c(u1, u2, t1, t2, r)}

\item{add}{par}

\item{s.point}{par}

\item{...}{par}
}
\value{
plot
}
\description{
ROC plot
}
\examples{
sROC(c(1,1, 0.5, 0.5, -0.6))

}
