% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sroc.R
\name{sROC}
\alias{sROC}
\title{Plot a summary ROC plot for dtametasa object}
\usage{
sROC(
  object,
  add = FALSE,
  sroc.col = 1,
  sroc.lty = 1,
  sroc.lwd = 1,
  xlab = "1-specificity",
  ylab = "Sensitivity",
  plot.ci = TRUE,
  ci.level = 0.95,
  sroc.ci.col = "grey48",
  sroc.ci.lwd = 1,
  sroc.ci.lty = 2,
  add.spoint = TRUE,
  spoint.pch = 18,
  spoint.col = 1,
  spoint.cex = 2,
  ...
)
}
\arguments{
\item{object}{The object from function \code{dtametasa.fc} or \code{dtametasa.rc};
or a vector of \code{c(u1, u2, t1, t2, r)}}

\item{add}{Whether to add the plot into an existed plot.
Default is \code{FALSE}, to create a new plot.}

\item{sroc.col}{The color of sROC.
Default is black.}

\item{sroc.lty}{The line type of sROC.
Default is solid.}

\item{sroc.lwd}{The line width of sROC.
Default is 1, solid.}

\item{xlab}{The label of x-axis.
Default is: 1-specificity.}

\item{ylab}{The label of y-axis.
Default is Sensitivity.}

\item{plot.ci}{Whether to plot confidence interval of sROC.
Default is \code{TRUE}, to plot the CI.}

\item{ci.level}{The significance level of confidence interval of sROC.
Default is \code{0.95}}

\item{sroc.ci.col}{The color of confidence interval of sROC.
Default is grey.}

\item{sroc.ci.lwd}{The line width of confidence interval of sROC.
Default is 1.}

\item{sroc.ci.lty}{The line type of confidence interval of sROC.
Default is 2, dashed.}

\item{add.spoint}{Whether to add the summary point in the sROC plot.
Default it not the add.}

\item{spoint.pch}{The type of the point.
Default is 19.}

\item{spoint.col}{The color of the point.
Default is black.}

\item{spoint.cex}{The size of the point.
Default is 1.}

\item{...}{Other augments in function \code{\link{points}} or function \code{\link{curve}}}
}
\value{
sROC plot
}
\description{
Plot a single sROC plot
}
\examples{

sa1 <- dtametasa.fc(IVD, p = 0.7)
sROC(sa1)

}
\seealso{
\code{\link[graphics]{points}},
\code{\link[graphics]{curve}},
\code{\link{dtametasa.fc}},
\code{\link{dtametasa.rc}}.
}
