% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sROC.R
\name{sROC}
\alias{sROC}
\title{Plot a single summary ROC plot}
\usage{
sROC(
  object,
  add = FALSE,
  roc.col = 1,
  roc.lty = 1,
  roc.lwd = 1,
  add.sum.point = TRUE,
  spoint.pch = 19,
  spoint.col = 1,
  spoint.cex = 1,
  xlab = "1-specificity",
  ylab = "Sensitivity",
  ...
)
}
\arguments{
\item{object}{The object from function \code{dtasens1} or \code{dtasens2};
or a vector of \code{c(u1, u2, t1, t2, r)}}

\item{add}{Whether to add the plot into an existed plot.
Default is \code{FALSE}, to create a new plot.}

\item{roc.col}{The color of sROC.
Default is black.}

\item{roc.lty}{The line type of sROC.
Default is solid.}

\item{roc.lwd}{The line width of sROC.
Default is 1.}

\item{add.sum.point}{Whether to add the summary point in the sROC plot.
Default it not the add.}

\item{spoint.pch}{The type of the point.
Default is 19.}

\item{spoint.col}{The color of the point.
Default is black.}

\item{spoint.cex}{The size of the point.
Default is 1.}

\item{xlab}{The label of x-axis.
Default is: 1-specificity.}

\item{ylab}{The label of y-axis.
Default is Sensitivity.}

\item{...}{Other augments in function \code{\link{points}} or function \code{\link{curve}}}
}
\value{
sROC plot
}
\description{
Plot a single sROC plot
}
\examples{

opt1 <- dtasens1(IVD, p = 0.7)
sROC(opt1)

par.vec <- c(1,1, 0.5, 0.5, -0.6)
sROC(par.vec)

}
\seealso{
\code{\link[graphics]{points}},
\code{\link[graphics]{curve}},
\code{\link{dtasens1}},
\code{\link{dtasens2}}.
}
