% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sROC.R
\name{sROC}
\alias{sROC}
\title{Single ROC plot}
\usage{
sROC(
  object,
  add = FALSE,
  roc.col = 1,
  roc.lty = 1,
  roc.lwd = 1,
  add.sum.point = TRUE,
  spoint.pch = 19,
  spoint.col = 1,
  spoint.cex = 1,
  ...
)
}
\arguments{
\item{object}{c(u1, u2, t1, t2, r)}

\item{add}{par}

\item{roc.col}{par}

\item{roc.lty}{par}

\item{roc.lwd}{par}

\item{add.sum.point}{par}

\item{spoint.pch}{par}

\item{spoint.col}{par}

\item{spoint.cex}{par}

\item{...}{par}
}
\value{
plot
}
\description{
Single ROC plot
}
\examples{

par.vec <- c(1,1, 0.5, 0.5, -0.6)
sROC(par.vec)

opt1 <- dtasens1(IVD, p = 0.7)
sROC(opt1)

}
