% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sROC.R
\name{SROC}
\alias{SROC}
\title{Summary ROC curves}
\usage{
SROC(
  object,
  sroc.type = c("sroc", "hsroc"),
  sroc.cols = gray.colors(ncol(as.matrix(object)), gamma = 1, start = 0, end = 0.5),
  sroc.lty = 1,
  sroc.lwd = 1,
  add.spoint = TRUE,
  sp.pch = 19,
  sp.cex = 1,
  plot.ci = FALSE,
  sroc.ci.col = "red",
  sroc.ci.lty = 2,
  sroc.ci.lwd = 1,
  ci.level = 0.95,
  xlab = "FPR",
  ylab = "TPR",
  addon = FALSE,
  ...
)
}
\arguments{
\item{object}{(1) The result from function \code{dtametasa.fc} or \code{dtametasa.rc} (single curve);
(2) a matrix of which the rows are \code{c(u1, u2, t1, t2, r)} (multiple curves);
(3) a vector of \code{c(u1, u2, t1, t2, r)} (single curve).}

\item{sroc.type}{Plot Reitsma's SROC curve(\code{"sroc"}) or Rutter's HSROC curve (\code{"hsroc"}),}

\item{sroc.cols}{Set a vector of colors for single/multiple SROC curves.
Default uses grey's colors.}

\item{sroc.lty}{Line type of SROC curve.}

\item{sroc.lwd}{Line width of SROC curve.}

\item{add.spoint}{Whether to add the summary point on the SROC curve.}

\item{sp.pch}{Type of the summary point.}

\item{sp.cex}{Size of the summary point.}

\item{plot.ci}{Whether to plot the confidence intervals of the SROC curve.
It is only for the result from function \code{dtametasa.fc} or \code{dtametasa.rc}}

\item{sroc.ci.col}{Colors of the confidence intervals.}

\item{sroc.ci.lty}{Line type of the confidence intervals.}

\item{sroc.ci.lwd}{Line width of the confidence intervals.}

\item{ci.level}{Significant level of the confidence intervals.}

\item{xlab}{Label of x-axis.}

\item{ylab}{Label of y-axis.}

\item{addon}{Whether to add the SROC curves onto an existed plot.
Default is \code{FALSE}, to start from a new plot.}

\item{...}{Other augments in function \code{\link{plot.default}} or function \code{\link{curve}}}
}
\value{
SROC curves
}
\description{
Plot the SROC curves
}
\examples{
p.seq <- seq(1, 0.5, -0.1)
sa1.seq <- sapply(p.seq, function (p) dtametasa.fc(IVD, p)$par)

SROC(sa1.seq, sroc.type = "sroc")
SROC(sa1.seq, sroc.type = "hsroc")

sa.fit <- dtametasa.fc(IVD, p = 0.7, sauc.type = "sroc")
sa.fit
SROC(sa.fit, sroc.type = "sroc", pch = 19, plot.ci = TRUE)
SROC(sa.fit, sroc.type = "hsroc", pch = 19, plot.ci = TRUE)

}
\seealso{
\code{\link[graphics]{plot.default}},
\code{\link[graphics]{points}},
\code{\link[graphics]{curve}},
\code{\link{dtametasa.fc}},
\code{\link{dtametasa.rc}}.
}
